+++ lib/features/booking/domain/entitiy/booking.dart
@@ -11,12 +11,13 @@ const bookingHistoryStatus = [
 
 class BookingEntity {
   final String bookingId;
+
   /// date when booking is recorded, in local time
   final DateTime bookingDate;
   final String title;
   final BookingStatus bookingStatus;
   final String? bookingNote;
-  final int? rating;
+  final double? rating;
   final String? location;
   final String? imgUrl;
 
@@ -64,7 +65,7 @@ class BookingDetailEntity {
   final String? adminMessage;
   final String? note;
   final String? title;
-  final int? rating;
+  final double? rating;
   final String? location;
   final String? thumbnail;
